DB

Description

Within this module DB access functions allowing to query and to store the data are provided.

Initialising DB

initialise :: IO ()

The initialise function has a purpose to initialise database and 2 tables - companies list and stock prices companies table consists of company name, date downloaded and stored in DB and username who downloaded it stock prices table consists of company name, dates and corresponding volume and a range of stock prices for each date

Query functions

queryStock

Arguments

:: String
The String argument to provide the company name to query

-> IO ()

The queryStock has a purpose to query stock prices

queryDate

Arguments

:: String
The String argument to provide the company name to query

-> String
The String argument to provide the date to query

-> IO ()

The queryDate function has a purpose to get company information at a certain date (volume, open, close prices)

queryMinPrice :: IO ()

The queryMinPrice function has a purpose to get the lowest stock close price among the companies stored in DB (for each company only 100 rows are stored)

queryMaxPrice :: IO ()

The queryMaxPrice function has a purpose to get the highest stock close price among the companies stored in DB (for each company only 100 rows are stored)

convRowP :: [SqlValue] -> String

The convRowP function is a helper function to process prices rows to get close price

queryMaxVolume :: IO ()

The queryMaxVolume function has a purpose to get the highest stock volume among the companies stored in DB (for each company only 100 rows are stored)

convRowV :: [SqlValue] -> String

The convRowV is a helper function to process prices rows to get volume

queryCompanies :: IO ()

The queryCompanies function has a purpose to get companies list

convRow :: [SqlValue] -> String

The convRow function is a helper function to process prices rows

queryCompany

Arguments

:: String
The String argument to provide the company name to query

-> IO String

The queryCompany function has a purpose to get a company (to check if it is present in DB)

convRowC :: [SqlValue] -> String

The convRowC function is a helper function to process companies rows

Insert functions

insert

Arguments

:: Price
The Price argument to provide the prices to be inserted

-> IO ()

The insert function has a purpose to insert stock prices

insertMany

Arguments

:: String
The String argument to provide the company name to be inserted

-> [Price]
The [Price] argument to provide the prices to be inserted

-> IO ()

The insertMany function has a purpose to insert multiple rows of stock prices

insertCompany

Arguments

:: String
The String argument to provide the company name to be inserted

-> String
The String argument to provide the user name to be inserted

-> IO ()

The insertCompany function has a purpose to insert new company into companies list

type Date = (Integer, Int, Int)

The type synonym Date represents date format in a more understandable way

Helper functions

getYear

Arguments

:: Date
The Date argument to provide the date to be parsed

-> String
The String argument to return a year

The getYear function retrieves a year from date datatype

getMonth

Arguments

:: Date
The Date argument to provide the date to be parsed

-> String
The String argument to return a month

The getMonth function retrieves a month from date datatype

getDay

Arguments

:: Date
The Date argument to provide the date to be parsed

-> String
The String argument to return a day

The getDay function retrieves a day from date datatype

formRows

Arguments

:: String
The String argument to represent company name

-> [Price]
The [Price] argument to represent company prices

-> [[SqlValue]]

The formRows function is a helper function to generate rows from stock prices data

formRow

Arguments

:: String
The String argument to represent company name

-> Price
The Price argument to represent company prices

-> [SqlValue]

The formRow function is a helper function to generate a row from stock prices data