DataType

Description

Within this module the main DataType used for the program is provided together with its sample data.

DataType

data Price

Price constructor stores stock prices in a form of a Haskell datatype

Constructors

Price

Fields

date :: String

open :: String

high :: String

low :: String

close :: String

volume :: Int

adj_close :: String

Sample data

price1 :: Price

Price{date = "1991-08-20",open = "23",high = "25",low = "22", close = "24",volume = 1,adj_close = "12"}

price2 :: Price

Price{date = "1991-09-20",open = "23",high = "25",low = "22", close = "24",volume = 1,adj_close = "12"}

prices :: [Price]

Sample price list