Main

Description

Within this module the main interactive functionality is provided. This module is linked with: DataType, DB, NW and Parse

Interaction

main :: IO ()

The main function is the principle point of interaction with the user. ghc Main.hs will build the source together, if all the modules are in the same folder. Use ./Main to run the program

interactWith

Arguments

:: a
The a argument is used to pass a parsing function

-> IO ()

The interactWith function provides interaction with user

showIntro

Arguments

:: IO String
The IO(String) argument is used to return a username

The showIntro function shows introduction and ask for user name

showMenu

Arguments

:: String
The String argument is used to pass a username

-> IO ()

The showMenu function shows menu options and run menu until user exits

Downloading data

downloadPrices

Arguments

:: String
The String argument is used to pass a username

-> IO ()

The downloadPrices function downloads stock prices for a specified company

Query functions

listCompanies :: IO ()

The listCompanies function shows companies list from DB

getPrices :: IO ()

The getPrices function gets stock prices from DB for a specified company

runQuery :: IO ()

The runQuery function shows available to user queries

Exit function

exit

Arguments

:: String
The String argument is used to pass a username

-> IO ()

The exit function exits the program upon user choice of Exit option

Helper functions

getFromRight

Arguments

:: Either a b
The Either argument is used to enable error handling

-> b
The b argument is used to return a value

The getFromRight function parses Either format to get a value