Parse

Description

Within this module the functions to parse downloaded data into Haskell accessible data are provided.

Functions to prepare input

f :: String -> String

The f function splits data from web page into lines and pass it to fn

fn :: [String] -> [Price]

The fn function returns [] if no lines received, otherwise processes the lines

process :: [String] -> [Price]

The process function has a purpose to process data from web page

take100 :: (Num a, Eq a) => [a1] -> [a1] -> a -> [a1]

The take100 function takes first 100 lines of the document

Functions to parse input

parse :: [String] -> [Price]

The parse function parses the whole document

parseLine :: String -> Price

The parseLine function parses each line of .csv file

parseDate

Arguments

:: String
Text

-> String
Date

The parseDate function parses date

parseOpen

Arguments

:: String
Text

-> String
Open price

The parseOpen function parses open price

parseHigh

Arguments

:: String
Text

-> String
High price

The parseHigh function parses high price

parseLow

Arguments

:: String
Text

-> String
Low price

The parseLow function parses low price

parseClose

Arguments

:: String
Text

-> String
Close price

The parseClose function parse close price

parseVolume

Arguments

:: String
Text

-> Int
Stock volume

The parseVolume function parse stock volume

parseAdjCl

Arguments

:: String
Text

-> String
Adj Close

The parseAdjCl function parse adj close price

Helper functions

getRest

Arguments

:: String
Text

-> String
Drop anything that is already parsed

The getRest function is a helper function to parse a line